[fix] pycodestyleのフォーマットに近づけた
