+++ 10_add_if_else_and_stack_control_and_loop/test_element.py
@@ -3,7 +3,9 @@
 # Created by devel on 2018/11/08.
 from element import *
 import pytest
+
 from eval import to_char_gen
+
 
 def test_elem_number():
     expect_elem = Element(Etype.NUMBER, value=123)
